1st-order consistency at boundaries for the pressure-gradient term.    BE_momentum. SASPH. Body_Transport. Input.    The renormalization matrix is inverted after all the contributions are       collected.       Extraction of the new program unit «B_ren_gradp_inversion» from the          program unit «CalcVarLength».       The renormalization matrix for the pressure-gradient term is inverted just          after all its components are collected and just before the 1st-order          consistency scheme applies to all the inner and boundary terms of the          pressure-gradient term in the momentum equation. Program units          «AddBoundaryContributions_to_ME3D» and          «AddBoundaryContributions_to_ME2D». This also applies in the absence of          neighbouring SASPH frontiers to fasten the algorithm under general          conditions.    The fluid-fluid contributions to the momentum equation are postponed after       the boundary contributions to the momentum equation to effectively       include the boundary contributions of the renormalization matrix.    The matrix product for the renormalization of the fluid-fluid interactions       only applies after all the neighbour-dependent contributions are       collected (program unit “inter_EqMoto”).    Split the PPST (Pressure Particle Shifting Technique) term from the       pressure-gradient term (fluid-fluid interactions; “fluid - SASPH       frontier” interactions; fluid-body interactions).    Collection of the boundary terms to the inverse of the renormalization       matrix: body particles; 3D SASPH frontiers; 2D SASPH frontiers.    Renormalization of boundary contributions to the pressure-gradient term:       body particles; 3D SASPH frontiers; 2D SASPH frontiers. The matrix       product applies after all the neighbour-dependent contributions are       collected, but the body-particle contributions, due to algorithm needs.    Minimum threshold on the absolute value of the determinant for the inversion       of the renormalization matrix is reduced from 0.1 to 0.01.

+++ src/Post_processing/Print_Results.f90
@@ -360,8 +360,8 @@ if (nag>0) then
                      Domain%t_pre_iter) * 1.d0
       if (elapsed_time<0.d0) elapsed_time =                                    &
          elapsed_time + 366.d0 * 24.d0 * 60.d0 * 60.d0
-      aux_scal = int(elapsed_time / 3600.d0)
-      aux_scal_2 = int(aux_scal / 24.d0)
+      aux_scal = elapsed_time / 3600.d0
+      aux_scal_2 = aux_scal / 24.d0
       write(ulog,'(a,3(ES15.6,a))') "Elapsed time: ",                          &
          elapsed_time," s = ",aux_scal,                                        &
          " hours = ",aux_scal_2," days."
